requirements.txt
Flask==3.0.1
spotipy==2.23.0
flask_cors==4.0.0
flask_session==0.6.0
flask_bcrypt==1.0.1
python-dotenv==1.0.1
flask_sqlalchemy==3.1.1